refactor: move components to new dir
