+++ apps/client/src/components/ErrorPage.tsx
